VLM-RL/requirements.txt
carla==0.9.13
networkx
pygame
gym==0.23.0
gymnasium==0.28.1
stable-baselines3==2.0.0
open_clip_torch==2.20.0
opencv-python==4.7.0.68
# torch==1.13.1+cu116
shimmy==1.3.0
tensorboard
python-box
scikit-learn
h5py
xmltodict